+++ common/src/main/java/de/maxhenkel/voicechat/gui/audiodevice/SelectMicrophoneScreen.java
@@ -1,5 +1,6 @@
 package de.maxhenkel.voicechat.gui.audiodevice;
 
+import de.maxhenkel.configbuilder.entry.ConfigEntry;
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.voice.client.ClientManager;
@@ -17,9 +18,9 @@ import java.util.List;
 
 public class SelectMicrophoneScreen extends SelectDeviceScreen {
 
-    protected static final ResourceLocation MICROPHONE_ICON = new ResourceLocation(Voicechat.MODID, "textures/icons/microphone.png");
-    protected static final ITextComponent TITLE = new TextComponentTranslation("gui.voicechat.select_microphone.title");
-    protected static final ITextComponent NO_MICROPHONE = new TextComponentTranslation("message.voicechat.no_microphone").setStyle(new Style().setColor(TextFormatting.GRAY));
+    public static final ResourceLocation MICROPHONE_ICON = new ResourceLocation(Voicechat.MODID, "textures/icons/microphone.png");
+    public static final ITextComponent TITLE = new TextComponentTranslation("gui.voicechat.select_microphone.title");
+    public static final ITextComponent NO_MICROPHONE = new TextComponentTranslation("message.voicechat.no_microphone").setStyle(new Style().setColor(TextFormatting.GRAY));
 
     public SelectMicrophoneScreen(@Nullable GuiScreen parent) {
         super(TITLE, parent);
@@ -31,12 +32,7 @@ public class SelectMicrophoneScreen extends SelectDeviceScreen {
     }
 
     @Override
-    public String getSelectedDevice() {
-        return VoicechatClient.CLIENT_CONFIG.microphone.get();
-    }
-
-    @Override
-    public ResourceLocation getIcon(String device) {
+    public ResourceLocation getIcon() {
         return MICROPHONE_ICON;
     }
 
@@ -46,17 +42,7 @@ public class SelectMicrophoneScreen extends SelectDeviceScreen {
     }
 
     @Override
-    public String getVisibleName(String device) {
-        // return SoundManager.cleanDeviceName(device);
-        return device;
-    }
-
-    @Override
-    public void onSelect(String device) {
-        VoicechatClient.CLIENT_CONFIG.microphone.set(device).save();
-        ClientVoicechat client = ClientManager.getClient();
-        if (client != null) {
-            client.reloadAudio();
-        }
+    public ConfigEntry<String> getConfigEntry() {
+        return VoicechatClient.CLIENT_CONFIG.microphone;
     }
 }