+++ common/src/main/java/de/maxhenkel/voicechat/gui/widgets/MicActivationButton.java
@@ -5,15 +5,17 @@ import de.maxhenkel.voicechat.voice.client.MicrophoneActivationType;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TextComponentTranslation;
 
+import java.util.function.Consumer;
+
 public class MicActivationButton extends EnumButton<MicrophoneActivationType> {
 
-    private final VoiceActivationSlider voiceActivationSlider;
+    protected Consumer<MicrophoneActivationType> onChange;
 
-    public MicActivationButton(int id, int xIn, int yIn, int widthIn, int heightIn, VoiceActivationSlider voiceActivationSlider) {
+    public MicActivationButton(int id, int xIn, int yIn, int widthIn, int heightIn, Consumer<MicrophoneActivationType> onChange) {
         super(id, xIn, yIn, widthIn, heightIn, VoicechatClient.CLIENT_CONFIG.microphoneActivationType);
-        this.voiceActivationSlider = voiceActivationSlider;
+        this.onChange = onChange;
         updateText();
-        setVisibility();
+        onChange.accept(entry.get());
     }
 
     @Override
@@ -23,11 +25,7 @@ public class MicActivationButton extends EnumButton<MicrophoneActivationType> {
 
     @Override
     protected void onUpdate(MicrophoneActivationType type) {
-        setVisibility();
-    }
-
-    private void setVisibility() {
-        voiceActivationSlider.visible = MicrophoneActivationType.VOICE.equals(entry.get());
+        onChange.accept(type);
     }
 
 }