+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/KeyEvents.java
@@ -5,6 +5,7 @@ import de.maxhenkel.voicechat.gui.VoiceChatScreen;
 import de.maxhenkel.voicechat.gui.VoiceChatSettingsScreen;
 import de.maxhenkel.voicechat.gui.group.GroupScreen;
 import de.maxhenkel.voicechat.gui.group.JoinGroupScreen;
+import de.maxhenkel.voicechat.gui.onboarding.OnboardingManager;
 import de.maxhenkel.voicechat.gui.volume.AdjustVolumesScreen;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 import de.maxhenkel.voicechat.voice.common.ClientGroup;
@@ -29,6 +30,8 @@ public class KeyEvents {
     public static KeyBinding KEY_TOGGLE_RECORDING;
     public static KeyBinding KEY_ADJUST_VOLUMES;
 
+    public static KeyMapping[] ALL_KEYS;
+
     public KeyEvents() {
         minecraft = Minecraft.getMinecraft();
         ClientCompatibilityManager.INSTANCE.onHandleKeyBinds(this::handleKeybinds);
@@ -43,14 +46,33 @@ public class KeyEvents {
         KEY_GROUP = ClientCompatibilityManager.INSTANCE.registerKeyBinding(new KeyBinding("key.voice_chat_group", Keyboard.KEY_G, "key.categories.voicechat"));
         KEY_TOGGLE_RECORDING = ClientCompatibilityManager.INSTANCE.registerKeyBinding(new KeyBinding("key.voice_chat_toggle_recording", Keyboard.KEY_NONE, "key.categories.voicechat"));
         KEY_ADJUST_VOLUMES = ClientCompatibilityManager.INSTANCE.registerKeyBinding(new KeyBinding("key.voice_chat_adjust_volumes", Keyboard.KEY_NONE, "key.categories.voicechat"));
+
+        ALL_KEYS = new KeyMapping[]{
+                KEY_PTT, KEY_WHISPER, KEY_MUTE, KEY_DISABLE, KEY_HIDE_ICONS, KEY_VOICE_CHAT, KEY_VOICE_CHAT_SETTINGS, KEY_GROUP, KEY_TOGGLE_RECORDING, KEY_ADJUST_VOLUMES
+        };
     }
 
     private void handleKeybinds() {
+        if (OnboardingManager.isOnboarding()) {
+            for (KeyMapping allKey : ALL_KEYS) {
+                if (allKey.consumeClick()) {
+                    OnboardingManager.startOnboarding(null);
+                    return;
+                }
+            }
+            return;
+        }
+
         ClientVoicechat client = ClientManager.getClient();
         ClientPlayerStateManager playerStateManager = ClientManager.getPlayerStateManager();
         if (KEY_VOICE_CHAT.isPressed()) {
             if (GuiScreen.isAltKeyDown()) {
-                ClientManager.getDebugOverlay().toggle();
+                if (Screen.hasControlDown()) {
+                    VoicechatClient.CLIENT_CONFIG.onboardingFinished.set(false).save();
+                    minecraft.player.displayClientMessage(Component.literal("Onboarding status has been reset"), true);
+                } else {
+                    ClientManager.getDebugOverlay().toggle();
+                }
             } else {
                 minecraft.displayGuiScreen(new VoiceChatScreen());
             }