+++ common/src/main/java/de/maxhenkel/voicechat/voice/client/RenderEvents.java
@@ -2,6 +2,7 @@ package de.maxhenkel.voicechat.voice.client;
 
 import de.maxhenkel.voicechat.Voicechat;
 import de.maxhenkel.voicechat.VoicechatClient;
+import de.maxhenkel.voicechat.gui.onboarding.OnboardingManager;
 import de.maxhenkel.voicechat.intercompatibility.ClientCompatibilityManager;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiScreen;
@@ -179,6 +180,9 @@ public class RenderEvents {
     }
 
     private boolean shouldShowIcons() {
+        if (OnboardingManager.isOnboarding()) {
+            return false;
+        }
         if (ClientManager.getClient() != null && ClientManager.getClient().getConnection() != null && ClientManager.getClient().getConnection().isInitialized()) {
             return true;
         }